+++ src/main/java/com/frejdh/util/common/functional/ThrowingFunction.java
@@ -1,0 +1,18 @@
+package com.frejdh.util.common.functional;
+
+import java.util.function.Function;
+
+@FunctionalInterface
+public interface ThrowingFunction<T, R> extends Function<T, R> {
+	@Override
+	default R apply(T parameter) {
+		try {
+			return applyThrows(parameter);
+		} catch (final Exception e) {
+			throw new RuntimeException(e);
+		}
+	}
+
+	R applyThrows(T parameter) throws Exception;
+}
+